[apps/web] refactor modals
