--- apps/web/components/Modal/index.module.scss
@@ -1,10 +1,0 @@
-.modal {
-  position: relative;
-  z-index: 1000;
-}
-
-.fullscreen {
-  border-radius: 0;
-  min-height: 100vh;
-  width: 100vh;
-}
