+++ packages/ui/src/Modal/index.module.scss
@@ -1,0 +1,91 @@
+@import '../colors.scss';
+
+.modal {
+  position: relative;
+  z-index: 1000;
+}
+
+.overlay {
+  background-color: rgba($color-gray-500, 0.75);
+  top: 0px;
+  right: 0px;
+  bottom: 0px;
+  left: 0px;
+  transition-property: opacity;
+  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+  transition-duration: 150ms;
+  opacity: 1;
+  position: fixed;
+}
+
+.container {
+  top: 0px;
+  right: 0px;
+  bottom: 0px;
+  left: 0px;
+  overflow-y: auto;
+  position: fixed;
+  z-index: 10;
+}
+
+.center {
+  align-items: center;
+  display: flex;
+  justify-content: center;
+  min-height: 100%;
+  padding: 0;
+}
+
+.hidden {
+  opacity: 0;
+}
+
+.visible {
+  opacity: 1;
+}
+
+.enter {
+  transition-timing-function: cubic-bezier(0, 0, 0.2, 1);
+  transition-duration: 300ms;
+}
+
+.leave {
+  transition-timing-function: cubic-bezier(0.4, 0, 1, 1);
+  transition-duration: 200ms;
+}
+
+.start {
+  opacity: 0;
+  transform: scaleX(0.95) scaleY(0.95);
+}
+
+.stop {
+  opacity: 1;
+  transform: scaleX(1) scaleY(1);
+}
+
+.content {
+  background: white;
+  border-radius: 0.5rem;
+  box-shadow: 0px 2px 8px -4px rgba($color-gray-700, 0.5);
+  overflow: hidden;
+  text-align: left;
+  transition-property: all;
+  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+  transition-duration: 150ms;
+  padding: 1rem;
+  position: relative;
+  @media (min-width: 640px) {
+    margin-bottom: 2rem;
+    margin-top: 2rem;
+    max-width: 24rem;
+    padding: 1.5rem;
+    width: 100%;
+  }
+}
+
+.fullscreen {
+  border-radius: 0;
+  min-height: 100vh;
+  width: 100vh;
+}
